requirements.txt
dj-rest-auth==2.2.2
Django==4.0.1
django-allauth==0.47.0
django-cors-headers==3.11.0
django-webpack-loader==1.4.1
djangorestframework==3.13.1
Pillow==9.0.0
